// src/models/public/ActionValues.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { ActionsActionId } from './Actions';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.action_values */
export type ActionValuesActionValueId = string & { __brand: 'public.action_values' };

/**
 * Represents the table public.action_values
 * value-ing actions. Measuring or assessing
 */
export default interface ActionValues {
  action_value_id: ActionValuesActionValueId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  action_id: ActionsActionId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer values */
  value_integer: number | null;

  /** Used for numeric values */
  value_numeric: number | null;

  /** Used for text values */
  value_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.action_values
 * value-ing actions. Measuring or assessing
 */
export interface ActionValuesInitializer {
  /** Default value: uuid_generate_v7() */
  action_value_id?: ActionValuesActionValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  action_id?: ActionsActionId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.action_values
 * value-ing actions. Measuring or assessing
 */
export interface ActionValuesMutator {
  action_value_id?: ActionValuesActionValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  action_id?: ActionsActionId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}